% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_int.R
\name{as_integer16_or_64}
\alias{as_integer16_or_64}
\title{as_integer16_or_64}
\usage{
as_integer16_or_64(x)
}
\arguments{
\item{x}{integerish vec}
}
\value{
int or int64
}
\description{
coerce to integer. if too large, coerces to 64-bit integer
}
