% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_setwd.R
\name{auto_setwd}
\alias{auto_setwd}
\title{auto setwd}
\usage{
auto_setwd()
}
\description{
Call from a saved R script. Automatically sets your working directory to the directory that you saved the current
R script in. Takes no arguments.
}
