% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_frame.R
\name{clean_frame}
\alias{clean_frame}
\title{Clean Data Frame}
\usage{
clean_frame(.data)
}
\arguments{
\item{.data}{a data frame}
}
\value{
data frame
}
\description{
Uses the functions of framecleaner and other operations to apply cleaning operations to a data frame
}
\examples{

iris \%>\%
clean_frame()
}
