% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_flag.R
\name{create_flag}
\alias{create_flag}
\title{create flag}
\usage{
create_flag(.data, col, flag, full_name = F, drop = F)
}
\arguments{
\item{.data}{data frame}

\item{col}{column}

\item{flag}{column entry}

\item{full_name}{Logical. default F. if T, new column name is original name + flag. other wise just flag}

\item{drop}{logical. default F. If T, drop original column.}
}
\value{
data frame
}
\description{
create flag
}
