% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_dates.R
\name{date_yq}
\alias{date_yq}
\title{date_yq}
\usage{
date_yq(x)
}
\arguments{
\item{x}{a date}
}
\value{
date vector
}
\description{
creates a quarterly date vector from a date vector
}
\examples{

seq.Date(lubridate::ymd(20200101), lubridate::ymd(20220101), length.out = 10) -> d1
d1 \%>\%
  tibble::enframe() \%>\%
  dplyr::mutate(YQ = date_yq(value))
}
