% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{fct_or_prob}
\alias{fct_or_prob}
\title{fct_or_prob}
\usage{
fct_or_prob(x, first_level = NULL, order_fct = FALSE)
}
\arguments{
\item{x}{vector}

\item{first_level}{character string to set the first level of the factor}

\item{order_fct}{logical. ordered factor?}
}
\value{
logical
}
\description{
fct_or_prob
}
\keyword{internal}
