% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_na.R
\name{fill_na}
\alias{fill_na}
\title{Fill NAs}
\usage{
fill_na(mdb, ..., fill = 0)
}
\arguments{
\item{mdb}{data frame}

\item{...}{tidyselect specification of columns}

\item{fill}{value to fill NAs}
}
\value{
data frame
}
\description{
use tidyselect to fill NA values
}
