% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_dir.R
\name{import_dir}
\alias{import_dir}
\title{import directory}
\usage{
import_dir(
  dir,
  ...,
  method = c("rio", "vroom", "vroom_jp", "read_csv"),
  return_type = c("df", "list")
)
}
\arguments{
\item{dir}{dir path}

\item{...}{arguments passed to import method}

\item{method}{import method chosen from import tibble}

\item{return_type}{default is to bind dataframes together and remove duplicates. only recommended
for a folder of files with the same data format. otherwise specify return as list of data frames}
}
\value{
data frame
}
\description{
import directory
}
