% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{is_integery}
\alias{is_integery}
\title{is integery}
\usage{
is_integery(x)
}
\arguments{
\item{x}{a vector}
}
\value{
logical
}
\description{
practical extension of rlang::is_integerish that supports integer64, integers in character strings,
and doesn't coerce factors
}
\keyword{internal}
