% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad_cols.R
\name{pad_auto}
\alias{pad_auto}
\title{pad auto}
\usage{
pad_auto(mdb, ..., side = "left", pad = "0")
}
\arguments{
\item{mdb}{data frame}

\item{...}{tidyselect specification}

\item{side}{str_pad side}

\item{pad}{str_pad pad}
}
\value{
data frame
}
\description{
Automatically pads elements of a column to the largest sized element. Useful when an integer code with
leading zeros is read in as an integer and needs to be fixed.
}
