% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad_cols.R
\name{pad_col}
\alias{pad_col}
\title{pad column}
\usage{
pad_col(mdb, col, width, pad = "0", side = "left")
}
\arguments{
\item{mdb}{data frame}

\item{col}{column}

\item{width}{str_pad width}

\item{pad}{str_pad pad}

\item{side}{str_pad side}
}
\value{
data frame
}
\description{
wrapper around mutate and str_pad
}
