% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relocate_all.R
\name{relocate_all}
\alias{relocate_all}
\title{Relocate All}
\usage{
relocate_all(.data)
}
\arguments{
\item{.data}{data frame}
}
\value{
data frame
}
\description{
Arranges columns alphabetically and then by type
}
