% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_whitespace.R
\name{remove_whitespace}
\alias{remove_whitespace}
\title{Remove Whitespace}
\usage{
remove_whitespace(.data, ...)
}
\arguments{
\item{.data}{data frame}

\item{...}{tidyselect specification (default selection:  all character columns)}
}
\value{
data frame
}
\description{
Remove whitespace from columns using a tidyselect specification.
}
\examples{

tibble::tibble(a = c(" a ", "b ", " c")) -> t1

t1

t1 \%>\%
remove_whitespace()
}
