% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{set_chr}
\alias{set_chr}
\title{set character}
\usage{
set_chr(.data, ...)
}
\arguments{
\item{.data}{dataframe}

\item{...}{tidyselect. Default selection: none}
}
\value{
tibble
}
\description{
set character
}
