% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_dates.R
\name{set_dates}
\alias{set_dates}
\title{Set Dates}
\usage{
set_dates(.data, ...)
}
\arguments{
\item{.data}{data frame}

\item{...}{a tidyselect specification}
}
\value{
data frame
}
\description{
Finds character columns in a data frame that represent dates and sets their type to date
}
