% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{set_dbl}
\alias{set_dbl}
\alias{set_dbl.character}
\alias{set_dbl.factor}
\alias{set_dbl.Date}
\alias{set_dbl.numeric}
\alias{set_dbl.data.frame}
\title{set double}
\usage{
set_dbl(.data, ...)

\method{set_dbl}{character}(x)

\method{set_dbl}{factor}(x)

\method{set_dbl}{Date}(x)

\method{set_dbl}{numeric}(x)

\method{set_dbl}{data.frame}(.data, ...)
}
\arguments{
\item{.data}{dataframe}

\item{...}{tidyselect. Default selection: none}
}
\value{
tibble
}
\description{
set double
}
\examples{



date_col <- c(lubridate::ymd(20180101), lubridate::ymd(20210420))


tibble::tibble(int = c(1L, 2L),
fct = factor(c(10, 11)),
date = date_col,
chr = c("a2.1", "rtg50.5")) -> t1

t1

t1 \%>\%
set_dbl(tidyselect::everything())

# s3 method works for vectors individually
# custom date coercion to represent date as a number. For lubridate's coercion method, use set_int
date_col \%>\%
set_dbl

}
