% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{set_dbl}
\alias{set_dbl}
\title{set double}
\usage{
set_dbl(.data, ...)
}
\arguments{
\item{.data}{dataframe}

\item{...}{tidyselect. Default selection: none}
}
\value{
tibble
}
\description{
set double
}
\examples{

tibble::tibble(a = c(1L, 2L), b = factor(c(10, 11))) \%>\%
set_dbl(c(a,b))

}
