% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{set_int}
\alias{set_int}
\alias{set_int.data.frame}
\alias{set_int.grouped_df}
\title{set integer}
\usage{
set_int(.data, ...)

\method{set_int}{data.frame}(.data, ...)

\method{set_int}{grouped_df}(.data, ...)
}
\arguments{
\item{.data}{dataframe}

\item{...}{tidyselect. Default Selecton: integerish doubles or integerish characters}
}
\value{
tibble
}
\description{
set integer
}
\examples{

# automaticaly turns "integerish" characters or doubles into int values

int_vec <- c("1", "2", "10")

tibble::tibble(
x = int_vec,
y = c(1.0, 5.0, 20.0)) -> tbl

# automatically coerce integerish cols in a tibble
tbl

tbl \%>\%
set_int()

# s3 method works for vectors as well

int_vec

int_vec \%>\%
set_int()
}
