% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_lgl.R
\name{set_lgl.data.frame}
\alias{set_lgl.data.frame}
\alias{set_lgl}
\alias{set_lgl.default}
\alias{set_lgl.numeric}
\alias{set_lgl.character}
\title{set logical}
\usage{
\method{set_lgl}{data.frame}(.data, ..., true_level = 1L)

set_lgl(.data, ..., true_level = 1L)

\method{set_lgl}{default}(.data, ...)

\method{set_lgl}{numeric}(.data, ..., true_level = 1L)

\method{set_lgl}{character}(.data, ..., true_level = c("T", "TRUE"))
}
\arguments{
\item{.data}{dataframe}

\item{...}{tidyselect. Default selection: none}

\item{true_level}{specify the value to set as TRUE. Default value is 1 for seamless conversion between logicals and integers. Can be given as a vector of values.}
}
\value{
dataframe
}
\description{
note: for non-binary data, all values other than the true_level will be set to false
}
\examples{

# convert a 1/0 vector back into T/F

tibble::tibble(x = c(1, 0, 0, 1, 0, 1)) \%>\%
set_lgl(x)
}
