% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{set_lgl}
\alias{set_lgl}
\title{set logical}
\usage{
set_lgl(.data, ..., true_level = 1)
}
\arguments{
\item{.data}{dataframe}

\item{...}{tidyselect}

\item{true_level}{specify the value to set as TRUE}
}
\value{
tibble
}
\description{
note: for non-binary data, all values other than the true_level will be set to false
}
