% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_na.R
\name{set_na}
\alias{set_na}
\title{Set NAs}
\usage{
set_na(.data, ..., vec = c("-", "", " ", "null"))
}
\arguments{
\item{.data}{data frame}

\item{...}{tidyselect specification}

\item{vec}{vector of elements to replace with NA}
}
\value{
data frame
}
\description{
Set elements to NA values using tidyselect specification.
Don't use this function on columns of different modes at once.
}
