% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{set_type_groups}
\alias{set_type_groups}
\title{set type groups}
\usage{
set_type_groups(.data, ..., setter)
}
\arguments{
\item{.data}{dataframe}

\item{...}{tidyselect. default selection is integerish doubles or integerish characters}

\item{setter}{which setter function to use}
}
\value{
tibble
}
\description{
set type groups
}
\keyword{internal}
