% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{vroom_jp}
\alias{vroom_jp}
\title{vroom jp}
\usage{
vroom_jp(path)
}
\arguments{
\item{path}{filepath}
}
\value{
a tibble
}
\description{
wrapper around vroom specifically to read in japanese csv files encoded in shift-jis
}
