% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_frame.R
\name{clean_frame}
\alias{clean_frame}
\title{Clean Data Frame}
\usage{
clean_frame(.data)
}
\arguments{
\item{.data}{a data frame}
}
\value{
data frame
}
\description{
Uses the functions of framecleaner and other operations to apply cleaning operations to a data frame
}
\details{
Functions applied in \code{clean_frame}

\itemize{
\item{\code{\link[janitor]{janitor::remove_empty}}}
\item{\code{\link[dplyr]{dplyr::rename_with}} \code{.fn = \link[base]{enc2utf8}}}
\item{\code{\link[janitor]{janitor::clean_names(case = "all_caps", ascii = FALSE)}}}
\item{\code{\link{set_int}}}
\item{\code{\link{set_date}}}
\item{\code{\link{make_na}}}
\item{\code{\link[tibble]{tibble::as_tibble}}}

}
}
\examples{

iris \%>\%
clean_frame()
}
