% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_na.R
\name{fill_na}
\alias{fill_na}
\title{Fill NAs}
\usage{
fill_na(.data, ..., fill = 0L, missing_type = c("all", "NA", "NaN", "Inf"))
}
\arguments{
\item{.data}{data frame}

\item{...}{tidyselect specification of columns}

\item{fill}{value to fill NAs}

\item{missing_type}{character vector. Choose what type of missing to fill. Default is all types.}
}
\value{
data frame
}
\description{
use tidyselect to fill NA values
Default behavior is to fill all integer or double columns cols with 0, preserving their types.
}
