% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_tibble.R
\name{import_tibble}
\alias{import_tibble}
\title{import tibble}
\usage{
import_tibble(path, ..., method = c("rio", "vroom", "vroom_jp", "read_csv"))
}
\arguments{
\item{path}{filepath}

\item{...}{other arguments}

\item{method}{method of import}
}
\value{
a tibble
}
\description{
wrapper around [rio::import()] to return a tibble instead of a data.table
This is the official file reader of TidyConsultant because of it's speed, accurate
parsing of filetypes, int64 support, and intelligent language parsing.
}
\details{
now supports multiple types of importing through [method]
}
