% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_tibble.R
\name{import_tibble}
\alias{import_tibble}
\title{import tibble}
\usage{
import_tibble(
  path,
  ...,
  method = c("vroom", "rio", "vroom_jp", "read_csv", "read_excel")
)
}
\arguments{
\item{path}{filepath}

\item{...}{other arguments}

\item{method}{method of import. default is rio}
}
\value{
a tibble
}
\description{
wrapper around multiple file readers. The default being [vroom] set to return a tibble, with [set_int] to encode integers.
Also available is rio and vroom_jp for japanese charaters.
}
\details{
Supports multiple types of importing through [method]
}
