% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_int.R
\name{is_integerish_character}
\alias{is_integerish_character}
\title{is integerish character}
\usage{
is_integerish_character(x)
}
\arguments{
\item{x}{a vector}
}
\value{
a logical
}
\description{
companion to rlang::is_integerish that returns true only for  integerish character vectors.
}
\keyword{internal}
