% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad_cols.R
\name{pad_col}
\alias{pad_col}
\title{pad column}
\usage{
pad_col(mdb, ..., width, pad = "0", side = "left")
}
\arguments{
\item{mdb}{data frame}

\item{...}{tidyselect}

\item{width}{str_pad width}

\item{pad}{str_pad pad}

\item{side}{str_pad side}
}
\value{
data frame
}
\description{
wrapper around mutate and str_pad
}
\examples{

 # manually pad with 0's (or other value)
 # use case over `pad_auto`: the desired width is greater than the widest element

 tibble::tibble(
 ID = c(2, 13, 86, 302)
 ) \%>\%
 pad_col(ID, width = 4)
}
