% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relocate_all.R
\name{relocate_all}
\alias{relocate_all}
\title{Relocate All}
\usage{
relocate_all(.data, ..., regex = NULL)
}
\arguments{
\item{.data}{data frame}

\item{...}{a tidyselect specification}
}
\value{
data frame
}
\description{
Arranges columns alphabetically and then by type
The user can supply a tidyselect argument to further specify relocations
}
