% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relocate_all.R
\name{relocate_all}
\alias{relocate_all}
\title{Relocate All}
\usage{
relocate_all(.data, ..., regex = NULL)
}
\arguments{
\item{.data}{data frame}

\item{...}{a tidyselect specification}

\item{regex}{a regular expression to match columns that will be put at the front of the df}
}
\value{
data frame
}
\description{
Arranges columns alphabetically and then by type
The user can supply a tidyselect argument to specify columns that should come first
}
\examples{

iris \%>\%
head \%>\%
relocate_all(matches("Petal"))

}
