% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_whitespace.R
\name{remove_whitespace}
\alias{remove_whitespace}
\title{Remove Whitespace}
\usage{
remove_whitespace(.data, ...)
}
\arguments{
\item{.data}{data frame}

\item{...}{tidyselect specification}
}
\value{
data frame
}
\description{
Remove whitespace from columns using a tidyselect specification.
}
