% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{set_int}
\alias{set_int}
\alias{set_int.data.frame}
\title{set integer}
\usage{
set_int(.data, ...)

\method{set_int}{data.frame}(.data, ...)
}
\arguments{
\item{.data}{dataframe}

\item{...}{tidyselect. Default Selecton: integerish doubles or integerish characters}
}
\value{
tibble
}
\description{
set integer
}
