% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{set_int}
\alias{set_int}
\title{set integer}
\usage{
set_int(.data, ...)
}
\arguments{
\item{.data}{dataframe}

\item{...}{tidyselect. default selection is integerish doubles or integerish characters}
}
\value{
tibble
}
\description{
set integer
}
