% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{set_lgl}
\alias{set_lgl}
\title{set logical}
\usage{
set_lgl(.data, ..., true_level = 1L)
}
\arguments{
\item{.data}{dataframe}

\item{...}{tidyselect. Default selection: none}

\item{true_level}{specify the value to set as TRUE. Default value is 1 for seamless conversion between logicals and integers.}
}
\value{
dataframe
}
\description{
note: for non-binary data, all values other than the true_level will be set to false
}
\examples{

# convert a 1/0 vector back into T/F

tibble::tibble(x = c(1, 0, 0, 1, 0, 1)) \%>\%
set_lgl(x)
}
